separate code properly
