+++ tokenizers/tk-encode/src/tokenizer/pipeline/mod.rs
@@ -1,7 +1,6 @@
 use std::iter::Enumerate;
 use std::sync::Arc;
 use std::vec::IntoIter;
-use std::{borrow::Cow, convert::TryFrom};
 
 #[cfg(feature = "normalizers")]
 use crate::normalizers::{
@@ -24,22 +23,7 @@ use crate::tokenizer::Model as _;
 use crate::{
     DecoderRuntime,
     models::bpe::{BpeScratch, PipelineBPE},
-    normalizers::{
-        byte_level::ByteLevel as ByteLevelNormalizer, metaspace::MetaspaceNormalizer,
-        prepend::Prepend, replace::Replace, strip::Strip, utils::Lowercase,
-    },
     pipeline::scratch_pool::{EncodeScratch, ScratchPool},
-    pre_tokenizers::{
-        bert::BertPreTokenizer,
-        delimiter::CharDelimiterSplit,
-        digits::Digits,
-        fixed_length::FixedLength,
-        punctuation::Punctuation,
-        sequence::PipelineSequence,
-        split::Split as SplitPretok,
-        whitespace::{Whitespace, WhitespaceSplit},
-    },
-    processors::template::{Piece, Sequence, Tokens},
     tokenizer::Decoder as _,
     vocab::bucket_added_vocabulary::AddedVocabulary as BucketAddedVocabulary,
 };
@@ -48,7 +32,7 @@ pub use parallel::PARALLEL_MIN_BYTES;
 #[cfg(feature = "parallelism")]
 use parallel::StreamingIter;
 
-use super::{Result, SplitDelimiterBehavior};
+use super::Result;
 
 #[cfg(feature = "parallelism")]
 mod parallel;
@@ -56,475 +40,19 @@ mod scratch_pool;
 
 pub use scratch_pool::ModelScratch;
 
-use atomsplit::classify::classify;
 pub use atomsplit::fsm::Span;
 
-pub trait Normalizer {
-    fn normalize<'a>(&self, input: &'a str) -> Result<Cow<'a, str>>;
-}
+mod normalizer;
+mod post_processor;
+mod pre_tokenizer;
 
-/// One normalization step of a [`PipelineTokenizer`].
-#[derive(Debug)]
-pub enum PipelineNormalizer {
-    /// This used to be a pretokenizer, it makes more sense as a normalizer.
-    Metaspace(MetaspaceNormalizer),
-    Replace(Replace),
-    Prepend(Prepend),
-    Strip(Strip),
-    Lowercase(Lowercase),
-    ByteLevel(ByteLevelNormalizer),
-    #[cfg(feature = "normalizers")]
-    Bert(BertNormalizer),
-    #[cfg(feature = "normalizers")]
-    StripAccents(StripAccents),
-    #[cfg(feature = "normalizers")]
-    NFC(NFC),
-    #[cfg(feature = "normalizers")]
-    NFD(NFD),
-    #[cfg(feature = "normalizers")]
-    NFKC(NFKC),
-    #[cfg(feature = "normalizers")]
-    NFKD(NFKD),
-    #[cfg(feature = "normalizers")]
-    Nmt(Nmt),
-    #[cfg(feature = "normalizers")]
-    Precompiled(PrecompiledNormalizer),
-}
+pub use normalizer::{Normalizer, NormalizerChain, PipelineNormalizer, normalize_all};
+pub use post_processor::{PipelinePostProcessor, Seq, Slice, Template, build_slices, compose};
+pub use pre_tokenizer::{
+    PipelinePreTokenizer, PreTokenizer, PreTokenizerScratch, SplitPolicy, split, split_delimiter,
+    split_matches,
+};
 
-// This replaces the previous normalizer sequence.
-pub struct NormalizerChain<'a>(pub &'a [PipelineNormalizer]);
-
-/// Runs `normalizers` in order, each one seeing what the one before it produced.
-///
-/// A normalizer returns a [`Cow`] (copy-on-write): a borrow when it had nothing to change, an owned
-/// `String` when it rewrote the text.
-///
-/// Chaining them needs care, because an owned `String` produced halfway through the chain is local to this function.
-/// The next normalizer may hand back a borrow of that locally owned [`String`], and that borrow cannot outlive the `String`.
-///
-/// Text no normalizer touches is never copied: it stays a borrow of `input` throughout.
-pub fn normalize_all<'a, N: Normalizer>(normalizers: &[N], input: &'a str) -> Result<Cow<'a, str>> {
-    let mut cow: Cow<'a, str> = Cow::Borrowed(input);
-    for normalizer in normalizers {
-        cow = match cow {
-            // Still `input` itself, which outlives us: pass it straight on.
-            Cow::Borrowed(s) => normalizer.normalize(s)?,
-            Cow::Owned(s) => {
-                let out = match normalizer.normalize(&s)? {
-                    // Rewritten again: keep the new `String`, drop ours.
-                    Cow::Owned(o) => Some(o),
-                    // Handed `s` back untouched: keep the `String` we already own.
-                    Cow::Borrowed(b) if b.as_ptr() == s.as_ptr() && b.len() == s.len() => None,
-                    // A borrow into `s` (for example, a substring of `s`): copy it out before `s` is dropped.
-                    Cow::Borrowed(b) => Some(b.to_owned()),
-                };
-                Cow::Owned(out.unwrap_or(s))
-            }
-        };
-    }
-    Ok(cow)
-}
-
-impl Normalizer for NormalizerChain<'_> {
-    fn normalize<'a>(&self, input: &'a str) -> Result<Cow<'a, str>> {
-        normalize_all(self.0, input)
-    }
-}
-
-impl Normalizer for PipelineNormalizer {
-    fn normalize<'a>(&self, input: &'a str) -> Result<Cow<'a, str>> {
-        match self {
-            Self::Metaspace(normalizer) => normalizer.normalize(input),
-            Self::Replace(normalizer) => normalizer.normalize(input),
-            Self::Prepend(normalizer) => normalizer.normalize(input),
-            Self::Strip(normalizer) => normalizer.normalize(input),
-            Self::Lowercase(normalizer) => normalizer.normalize(input),
-            Self::ByteLevel(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::Bert(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::StripAccents(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::NFC(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::NFD(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::NFKC(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::NFKD(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::Nmt(normalizer) => normalizer.normalize(input),
-            #[cfg(feature = "normalizers")]
-            Self::Precompiled(normalizer) => normalizer.normalize(input),
-        }
-    }
-}
-
-/// Range-based pre-tokenization: yields spans into the input rather than owned
-/// substrings, so the pipeline can pre-tokenize without allocating.
-///
-/// # Safety
-///
-/// [`PreTokenizer::pre_tokenize`] produces [`Span`] objects that are later consumed by [`PipelineTokenizer::encode_sequence`].
-/// For performance reasons, [`PipelineTokenizer::encode_sequence`] turns every span into a `&str` with [`str::get_unchecked`].
-/// As a consequence, every implementation of this trait *MUST* ensure the following invariants for each [`Span`]:
-///
-/// * `span.end <= text.len()`: the span is inside the text
-/// * `span.start <= span.end`: the span is not reversed
-/// * `text.is_char_boundary(start)` and `text.is_char_boundary(end)`: the start and end offset must be UTF-8 char boundaries
-/// * The offsets are relative to the `text`.
-pub unsafe trait PreTokenizer {
-    /// Split `text` into pre-tokens, appending to `out`. Ranges are into `text`.
-    /// `scratch` holds the working buffers, see [`PreTokenizerScratch`].
-    fn pre_tokenize(
-        &self,
-        text: &str,
-        scratch: &mut PreTokenizerScratch,
-        out: &mut Vec<Span>,
-    ) -> Result<()>;
-}
-
-/// The working buffers a [`PreTokenizer`] needs to split a text into pre-tokens.
-#[derive(Default)]
-pub struct PreTokenizerScratch {
-    /// One [`atomsplit`] atom tag per input byte, what [`atomsplit::classify::classify`] writes
-    /// and the FSMs read.
-    tags: Vec<u8>,
-    /// An intermediate buffer in which the FSM writes the Span before they get appended to the output
-    spans: Vec<Span>,
-    /// The two buffers a [`PipelineSequence`] reads and writes as each of its children subdivides
-    /// the spans the child before it produced. Handed out by [`Self::take_pair`] rather than
-    /// borrowed, so the sequence can pass the rest of the scratch to its children while it holds
-    /// them.
-    pair: [Vec<Span>; 2],
-}
-
-impl PreTokenizerScratch {
-    /// Tag every byte of `bytes` with its [`atomsplit`] atom class, run `fsm` over the tags, and
-    /// append the spans to `out`.
-    pub fn split_on_tags(
-        &mut self,
-        bytes: &[u8],
-        fsm: impl FnOnce(&[u8], &[u8], &mut [Span]) -> usize,
-        out: &mut Vec<Span>,
-    ) {
-        let n = bytes.len();
-        let Self { tags, spans, .. } = self;
-        if tags.len() < n {
-            tags.resize(n, 0);
-        }
-        if spans.len() < n + 1 {
-            spans.resize(n + 1, Span::default());
-        }
-        // Assign a tag to each byte
-        classify(bytes, &mut tags[..n]);
-        // Run the fsm on the tags to determine where to cut
-        let k = fsm(bytes, &tags[..n], &mut spans[..n + 1]);
-        // Copy spans to output
-        out.extend_from_slice(&spans[..k]);
-    }
-
-    /// Run `fsm` over `bytes` and append the spans it cut to `out`, for an FSM that scans the bytes
-    /// itself and so has no use for the atom tags. Skips the [`classify`]
-    pub fn split_on_bytes(
-        &mut self,
-        bytes: &[u8],
-        fsm: impl FnOnce(&[u8], &mut [Span]) -> usize,
-        out: &mut Vec<Span>,
-    ) {
-        let n = bytes.len();
-        let spans = &mut self.spans;
-        if spans.len() < n + 1 {
-            spans.resize(n + 1, Span::default());
-        }
-        let k = fsm(bytes, &mut spans[..n + 1]);
-        out.extend_from_slice(&spans[..k]);
-    }
-
-    /// Take the sequence's two buffers, leaving empty ones behind.
-    pub fn take_pair(&mut self) -> [Vec<Span>; 2] {
-        std::mem::take(&mut self.pair)
-    }
-
-    /// Give back what [`Self::take_pair`] handed out, so the next sequence reuses the allocations.
-    pub fn put_pair(&mut self, pair: [Vec<Span>; 2]) {
-        self.pair = pair;
-    }
-}
-
-/// The pre-tokenizers a [`PipelineTokenizer`] can run.
-#[allow(clippy::large_enum_variant)]
-#[derive(Debug, Clone, PartialEq)]
-pub enum PipelinePreTokenizer {
-    Bert(BertPreTokenizer),
-    Delimiter(CharDelimiterSplit),
-    Digits(Digits),
-    FixedLength(FixedLength),
-    Punctuation(Punctuation),
-    Sequence(PipelineSequence),
-    Split(SplitPretok),
-    #[cfg(feature = "unicode-scripts")]
-    UnicodeScripts(crate::pre_tokenizers::unicode_scripts::UnicodeScripts),
-    Whitespace(Whitespace),
-    WhitespaceSplit(WhitespaceSplit),
-    None,
-}
-
-// SAFETY: every arm but `None` forwards to another `PreTokenizer`, which upholds the contract itself.
-// `None` emits the single span covering all of `text`, whose ends are `0` and `text.len()`.
-unsafe impl PreTokenizer for PipelinePreTokenizer {
-    fn pre_tokenize(
-        &self,
-        text: &str,
-        scratch: &mut PreTokenizerScratch,
-        out: &mut Vec<Span>,
-    ) -> Result<()> {
-        match self {
-            Self::None => {
-                out.push(Span {
-                    start: 0,
-                    end: text.len() as u32,
-                });
-                Ok(())
-            }
-            Self::Bert(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Delimiter(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Digits(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::FixedLength(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Punctuation(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Sequence(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Split(pretok) => pretok.pre_tokenize(text, scratch, out),
-            #[cfg(feature = "unicode-scripts")]
-            Self::UnicodeScripts(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::Whitespace(pretok) => pretok.pre_tokenize(text, scratch, out),
-            Self::WhitespaceSplit(pretok) => pretok.pre_tokenize(text, scratch, out),
-        }
-    }
-}
-
-/// A post-processor compiled to a prefix and a suffix (slices of token IDs)
-///
-/// Example:
-/// ```text
-///     PipelinePostProcessor {
-///         prefix: vec![100].into_boxed_slice(),
-///         suffix: vec![101, 102].into_boxed_slice()
-///     };
-///
-///     [CLS] The quick Brown fox  [SEP]
-///     <100>|  <3> <4> <19> <67> | <101> <102>
-///   prefix |  sequence encoding | suffix
-/// ```
-///
-#[derive(Debug)]
-pub struct PipelinePostProcessor {
-    single: Template,
-    pair: Template,
-}
-
-impl PipelinePostProcessor {
-    /// Assemble one from an already-built single and pair template.
-    ///
-    /// Exists so both readers stop at the same door: the slim reader builds the two templates from
-    /// JSON, `tk-convert` builds them from a `PostProcessorWrapper`, and neither needs the
-    /// fields.
-    pub fn new(single: Template, pair: Template) -> Self {
-        Self { single, pair }
-    }
-
-    /// The two templates back out again, for composing a `Sequence` post-processor out of the
-    /// members it lowered to. Read-only: `compose` picks one of them, it does not edit them.
-    pub fn templates(&self) -> (&Template, &Template) {
-        (&self.single, &self.pair)
-    }
-}
-
-#[derive(Clone, Debug)]
-pub enum Slice {
-    Specials {
-        tokens: Box<[PipelineToken]>,
-        type_id: u8,
-    },
-    Sequence {
-        seq: Seq,
-        type_id: u8,
-    },
-}
-
-#[derive(Clone, Copy, Debug)]
-pub enum Seq {
-    A,
-    B,
-}
-
-#[derive(Debug)]
-pub struct Template {
-    slices: Box<[Slice]>,
-    n_special: usize,
-    has_type_ids: bool,
-}
-
-impl Template {
-    pub fn new(slices: Vec<Slice>) -> Self {
-        let n_special = slices
-            .iter()
-            .map(|s| {
-                if let Slice::Specials { tokens, .. } = s {
-                    tokens.len()
-                } else {
-                    0
-                }
-            })
-            .sum();
-        let has_type_ids = slices.iter().any(|s| match s {
-            Slice::Specials { type_id, .. } | Slice::Sequence { type_id, .. } => *type_id != 0,
-        });
-        Self {
-            slices: slices.into_boxed_slice(),
-            n_special,
-            has_type_ids,
-        }
-    }
-
-    pub fn slices(&self) -> &[Slice] {
-        &self.slices
-    }
-}
-
-// TODO: I don't think this is as optimized as can be yet, but we'll do that post v1.
-pub fn build_slices(pieces: &[Piece], specials: &Tokens, is_pair: bool) -> Result<Vec<Slice>> {
-    let (mut seen_a, mut seen_b) = (false, false);
-    let mut slices = Vec::new();
-    for piece in pieces {
-        match piece {
-            Piece::Sequence {
-                id: Sequence::A,
-                type_id,
-            } => {
-                if seen_a {
-                    return Err(
-                        "not supported: template references sequence A more than once".into(),
-                    );
-                }
-                seen_a = true;
-                slices.push(Slice::Sequence {
-                    seq: Seq::A,
-                    type_id: u8::try_from(*type_id)
-                        .map_err(|_| "not supported: type_id out of range")?,
-                });
-            }
-            Piece::Sequence {
-                id: Sequence::B,
-                type_id,
-            } => {
-                if seen_b {
-                    return Err(
-                        "not supported: template references sequence B more than once".into(),
-                    );
-                }
-                seen_b = true;
-                slices.push(Slice::Sequence {
-                    seq: Seq::B,
-                    type_id: u8::try_from(*type_id)
-                        .map_err(|_| "not supported: type_id out of range")?,
-                });
-            }
-            Piece::SpecialToken {
-                id: token_string,
-                type_id,
-            } => {
-                let special = specials.0.get(token_string).ok_or_else(|| {
-                    format!("not supported: unknown special token: `{token_string}`")
-                })?;
-                slices.push(Slice::Specials {
-                    tokens: special
-                        .ids()
-                        .iter()
-                        .map(|&id| PipelineToken::from(id))
-                        .collect(),
-                    type_id: u8::try_from(*type_id)
-                        .map_err(|_| "not supported: type_id out of range")?,
-                });
-            }
-        }
-    }
-    if !seen_a {
-        return Err("not supported: template does not reference sequence A".into());
-    }
-    if is_pair && !seen_b {
-        return Err("not supported: pair template does not reference sequence B".into());
-    }
-    if !is_pair && seen_b {
-        return Err(
-            "not supported: single template references sequence B (it should only refer to A)"
-                .into(),
-        );
-    }
-    Ok(slices)
-}
-
-/// A pass-through template does nothing: no special tokens, and sequences in the default
-/// arrangement (`$A`, or `$A $B` with the default type ids 0 then 1). Such a member is a no-op in a
-/// Sequence and is dropped when composing. Anything else adds tokens or reorders/retags.
-fn is_pass_through(slices: &[Slice]) -> bool {
-    matches!(
-        slices,
-        [Slice::Sequence {
-            seq: Seq::A,
-            type_id: 0
-        }] | [
-            Slice::Sequence {
-                seq: Seq::A,
-                type_id: 0
-            },
-            Slice::Sequence {
-                seq: Seq::B,
-                type_id: 1
-            }
-        ]
-    )
-}
-
-// TODO: this looks unused
-pub fn compose<'a>(templates: impl Iterator<Item = &'a Template>) -> Result<Template> {
-    let templates = templates.collect::<Vec<_>>();
-    let mut chosen: Option<&Template> = None;
-    for template in &templates {
-        if is_pass_through(&template.slices) {
-            continue;
-        }
-        if chosen.replace(template).is_some() {
-            return Err(
-                "post processor Sequence with multiple sequence referencing members is not supported".into(),
-            );
-        }
-    }
-    let chosen = chosen
-        .or_else(|| templates.first().copied())
-        .ok_or("empty Sequence post processor is not supported")?;
-    Ok(Template::new(chosen.slices.to_vec()))
-}
-
-impl Default for PipelinePostProcessor {
-    fn default() -> Self {
-        Self {
-            single: Template::new(vec![Slice::Sequence {
-                seq: Seq::A,
-                type_id: 0,
-            }]),
-            pair: Template::new(vec![
-                Slice::Sequence {
-                    seq: Seq::A,
-                    type_id: 0,
-                },
-                Slice::Sequence {
-                    seq: Seq::B,
-                    type_id: 1,
-                },
-            ]),
-        }
-    }
-}
 
 /// An output token. Carries only the vocabulary `id`, since offsets and the token
 /// string are dropped, which is all an encode-only caller needs.
@@ -1324,215 +852,6 @@ impl PipelineDecodeStream<'_> {
     }
 }
 
-/// What [`split`] does with each split it forms
-#[derive(Clone, Copy, PartialEq, Eq, Debug)]
-pub enum SplitPolicy {
-    /// Drop it, emit no split
-    Remove,
-    /// Emit it whole as one split
-    Keep,
-    /// Emit each character as its own split
-    Isolate,
-}
-
-/// Splits `text` into same-class groups, emitting each as a [`Span`]
-/// according to its [`SplitPolicy`].
-///
-/// `classify` maps each char to a small `Copy + Eq` class, the current
-/// split ends whenever the class changes (or on every char of an `Isolate`
-/// class), and `policy` decides what becomes of it. Ranges are byte offsets
-/// into `text`.
-#[inline(always)]
-pub fn split<C: Copy + PartialEq>(
-    text: &str,
-    out: &mut Vec<Span>,
-    classify: impl Fn(char) -> C,
-    policy: impl Fn(C) -> SplitPolicy,
-) {
-    let mut start: u32 = 0;
-    let mut prev: Option<C> = None;
-
-    for (i, ch) in text.char_indices() {
-        let c = classify(ch);
-        if let Some(p) = prev
-            && (p != c || policy(c) == SplitPolicy::Isolate)
-        {
-            if policy(p) != SplitPolicy::Remove {
-                out.push(Span {
-                    start,
-                    end: i as u32,
-                });
-            }
-            start = i as u32;
-        }
-        prev = Some(c);
-    }
-
-    if let Some(p) = prev
-        && policy(p) != SplitPolicy::Remove
-    {
-        out.push(Span {
-            start,
-            end: text.len() as u32,
-        });
-    }
-}
-
-/// Splits `text` around a single delimiter predicate, honoring the full
-/// [`SplitDelimiterBehavior`] contract. The pipeline-side equivalent of
-/// `NormalizedString::split(pattern, behavior)` for a char predicate.
-///
-/// The three non-merging behaviors reduce to a [`SplitPolicy`] on the delimiter
-/// class and reuse [`split`]. The two merge variants are their own single pass:
-/// - `MergedWithPrevious` cuts the split *after* each delimiter, so a delimiter
-///   joins the run before it (`"the-final"` -> `["the-", "final"]`).
-/// - `MergedWithNext` cuts *before* each delimiter, so it joins the run after it
-///   (`"the-final"` -> `["the", "-final"]`).
-pub fn split_delimiter(
-    text: &str,
-    out: &mut Vec<Span>,
-    is_delim: impl Fn(char) -> bool,
-    behavior: SplitDelimiterBehavior,
-) {
-    use SplitDelimiterBehavior::*;
-
-    let delim_policy = match behavior {
-        Removed => SplitPolicy::Remove,
-        Isolated => SplitPolicy::Isolate,
-        Contiguous => SplitPolicy::Keep,
-        MergedWithPrevious => {
-            let mut start: u32 = 0;
-            for (i, ch) in text.char_indices() {
-                if is_delim(ch) {
-                    let end = (i + ch.len_utf8()) as u32;
-                    out.push(Span { start, end });
-                    start = end;
-                }
-            }
-            if (start as usize) < text.len() {
-                out.push(Span {
-                    start,
-                    end: text.len() as u32,
-                });
-            }
-            return;
-        }
-        MergedWithNext => {
-            let mut start: u32 = 0;
-            for (i, ch) in text.char_indices() {
-                if is_delim(ch) {
-                    let i = i as u32;
-                    // skip the empty span before a leading run of delimiters
-                    if i > start {
-                        out.push(Span { start, end: i });
-                    }
-                    start = i;
-                }
-            }
-            if (start as usize) < text.len() {
-                out.push(Span {
-                    start,
-                    end: text.len() as u32,
-                });
-            }
-            return;
-        }
-    };
-
-    split(text, out, is_delim, |d| {
-        if d { delim_policy } else { SplitPolicy::Keep }
-    });
-}
-
-/// Applies a [`SplitDelimiterBehavior`] to a match segmentation and appends the
-/// resulting pieces to `out`.
-///
-/// `matches` is the `(offsets, is_match)` sequence covering the whole input,
-/// so regex matches interleaved with the gaps between them (exactly what
-/// `Pattern::find_matches` produces). This is the pipeline-side equivalent of
-/// the fold in `NormalizedString::split`; the arms mirror it exactly. Empty and
-/// removed pieces are dropped.
-pub fn split_matches(
-    out: &mut Vec<Span>,
-    matches: Vec<((usize, usize), bool)>,
-    behavior: SplitDelimiterBehavior,
-) {
-    use SplitDelimiterBehavior::*;
-
-    // (offsets, should_remove), mirroring `NormalizedString::split`.
-    let splits: Vec<((usize, usize), bool)> = match behavior {
-        Isolated => matches.into_iter().map(|(o, _)| (o, false)).collect(),
-        Removed => matches, // should_remove == is_match
-        Contiguous => {
-            let mut previous_match = false;
-            matches
-                .into_iter()
-                .fold(vec![], |mut acc, (offsets, is_match)| {
-                    if is_match == previous_match {
-                        if let Some(((_, end), _)) = acc.last_mut() {
-                            *end = offsets.1;
-                        } else {
-                            acc.push((offsets, false));
-                        }
-                    } else {
-                        acc.push((offsets, false));
-                    }
-                    previous_match = is_match;
-                    acc
-                })
-        }
-        MergedWithPrevious => {
-            let mut previous_match = false;
-            matches
-                .into_iter()
-                .fold(vec![], |mut acc, (offsets, is_match)| {
-                    if is_match && !previous_match {
-                        if let Some(((_, end), _)) = acc.last_mut() {
-                            *end = offsets.1;
-                        } else {
-                            acc.push((offsets, false));
-                        }
-                    } else {
-                        acc.push((offsets, false));
-                    }
-                    previous_match = is_match;
-                    acc
-                })
-        }
-        MergedWithNext => {
-            let mut previous_match = false;
-            let mut splits =
-                matches
-                    .into_iter()
-                    .rev()
-                    .fold(vec![], |mut acc, (offsets, is_match)| {
-                        if is_match && !previous_match {
-                            if let Some(((start, _), _)) = acc.last_mut() {
-                                *start = offsets.0;
-                            } else {
-                                acc.push((offsets, false));
-                            }
-                        } else {
-                            acc.push((offsets, false));
-                        }
-                        previous_match = is_match;
-                        acc
-                    });
-            splits.reverse();
-            splits
-        }
-    };
-
-    for ((start, end), should_remove) in splits {
-        if !should_remove && start != end {
-            out.push(Span {
-                start: start as u32,
-                end: end as u32,
-            });
-        }
-    }
-}
-
 pub trait Model {
     type Scratch: ModelScratch;
 
@@ -1640,149 +959,6 @@ impl Model for PipelineModel {
 #[cfg(test)]
 mod tests {
     use super::*;
-    use crate::pre_tokenizers::split::SplitPattern;
-    use crate::utils::byte_level::GPT2_REGEX_STR;
-
-    fn variant_name(pre_tokenizer: &PipelinePreTokenizer) -> &'static str {
-        match pre_tokenizer {
-            PipelinePreTokenizer::Bert(_) => "Bert",
-            PipelinePreTokenizer::Delimiter(_) => "Delimiter",
-            PipelinePreTokenizer::Digits(_) => "Digits",
-            PipelinePreTokenizer::FixedLength(_) => "FixedLength",
-            PipelinePreTokenizer::Punctuation(_) => "Punctuation",
-            PipelinePreTokenizer::Sequence(_) => "Sequence",
-            PipelinePreTokenizer::Split(_) => "Split",
-            #[cfg(feature = "unicode-scripts")]
-            PipelinePreTokenizer::UnicodeScripts(_) => "UnicodeScripts",
-            PipelinePreTokenizer::Whitespace(_) => "Whitespace",
-            PipelinePreTokenizer::WhitespaceSplit(_) => "WhitespaceSplit",
-            PipelinePreTokenizer::None => "None",
-        }
-    }
-
-    const HOSTILE: &[&str] = &[
-        "",
-        " ",
-        "\r",
-        "\r\n\n",
-        "hello world",
-        "café naïve",
-        "a\u{0301}b\u{0301}\u{0301}c",
-        "中文分词。ひらがな 한글",
-        "😀👍🏽 👨‍👩‍👧‍👦 x",
-        "\u{FEFF}leading bom",
-        "مرحبا\u{200F} العربية",
-        "नरेंद्र मोदी",
-        "hello▁world▁▁",
-        "a1b22c333d4444 42 ½²¼ Ⅷ",
-        "!!!...?! a, b; c",
-        "  \t  trailing   ",
-        "Ⓘ\u{200D}x_y a-b'c",
-    ];
-
-    /// Every variant must produce spans that [`PipelineTokenizer::encode_sequence`] can hand to `str::get_unchecked`
-    #[test]
-    fn every_pre_tokenizer_emits_sliceable_spans() {
-        use crate::pre_tokenizers::digits::Digits;
-        use crate::pre_tokenizers::fixed_length::FixedLength;
-        use crate::pre_tokenizers::punctuation::Punctuation;
-        use SplitDelimiterBehavior::*;
-
-        let literal_split = |pattern: &str, behavior| {
-            SplitPretok::new(SplitPattern::String(pattern.to_owned()), behavior, false).unwrap()
-        };
-        // The gpt2 regex is recognized, so this routes to `fsm_byte_level` with no regex backend.
-        let gpt2_split = SplitPretok::new(
-            SplitPattern::Regex(GPT2_REGEX_STR.to_owned()),
-            Isolated,
-            false,
-        )
-        .unwrap();
-
-        // `mut` + a cfg'd push rather than a `#[cfg]` on a `vec!` element: attributes on
-        // expression position are not stable.
-        #[allow(unused_mut)]
-        let mut cases = vec![
-            PipelinePreTokenizer::Bert(BertPreTokenizer),
-            PipelinePreTokenizer::Delimiter(CharDelimiterSplit::new(' ')),
-            PipelinePreTokenizer::Delimiter(CharDelimiterSplit::new('▁')),
-            PipelinePreTokenizer::Digits(Digits::new(true)),
-            PipelinePreTokenizer::Digits(Digits::new(false)),
-            PipelinePreTokenizer::FixedLength(FixedLength::new(3)),
-            PipelinePreTokenizer::FixedLength(FixedLength::new(0)),
-            PipelinePreTokenizer::Punctuation(Punctuation::new(Removed)),
-            PipelinePreTokenizer::Punctuation(Punctuation::new(Isolated)),
-            PipelinePreTokenizer::Punctuation(Punctuation::new(Contiguous)),
-            PipelinePreTokenizer::Punctuation(Punctuation::new(MergedWithPrevious)),
-            PipelinePreTokenizer::Punctuation(Punctuation::new(MergedWithNext)),
-            PipelinePreTokenizer::Sequence(PipelineSequence::new(vec![
-                PipelinePreTokenizer::WhitespaceSplit(WhitespaceSplit),
-                PipelinePreTokenizer::Punctuation(Punctuation::new(Isolated)),
-            ])),
-            PipelinePreTokenizer::Split(gpt2_split),
-            PipelinePreTokenizer::Split(literal_split("▁", MergedWithPrevious)),
-            PipelinePreTokenizer::Split(literal_split("▁", MergedWithNext)),
-            PipelinePreTokenizer::Split(literal_split(" ", Removed)),
-            PipelinePreTokenizer::Whitespace(Whitespace),
-            PipelinePreTokenizer::WhitespaceSplit(WhitespaceSplit),
-            PipelinePreTokenizer::None,
-        ];
-        #[cfg(feature = "unicode-scripts")]
-        cases.push(PipelinePreTokenizer::UnicodeScripts(
-            crate::pre_tokenizers::unicode_scripts::UnicodeScripts::new(),
-        ));
-
-        let mut covered: Vec<&str> = cases.iter().map(variant_name).collect();
-        covered.sort_unstable();
-        covered.dedup();
-        #[allow(unused_mut)]
-        let mut expected: Vec<&str> = vec![
-            "Bert",
-            "Delimiter",
-            "Digits",
-            "FixedLength",
-            "None",
-            "Punctuation",
-            "Sequence",
-            "Split",
-            "Whitespace",
-            "WhitespaceSplit",
-        ];
-        #[cfg(feature = "unicode-scripts")]
-        expected.push("UnicodeScripts");
-        expected.sort_unstable();
-        assert_eq!(
-            covered, expected,
-            "every PipelinePreTokenizer variant needs a case above",
-        );
-
-        let mut scratch = PreTokenizerScratch::default();
-        let mut spans = Vec::new();
-        for pre_tokenizer in &cases {
-            for text in HOSTILE {
-                spans.clear();
-                pre_tokenizer
-                    .pre_tokenize(text, &mut scratch, &mut spans)
-                    .unwrap();
-                for span in &spans {
-                    let range = span.range();
-                    assert!(
-                        range.start <= range.end,
-                        "{} reversed {span:?} on {text:?}",
-                        variant_name(pre_tokenizer),
-                    );
-                    assert!(
-                        text.is_char_boundary(range.start) && text.is_char_boundary(range.end),
-                        "{} cut {span:?} off a character boundary of {text:?}",
-                        variant_name(pre_tokenizer),
-                    );
-                    // What the pipeline does with the span. Slicing checked here is the point: it
-                    // panics on exactly the ranges `get_unchecked` would turn into undefined behavior.
-                    assert!(text.get(range).is_some());
-                }
-            }
-        }
-    }
 
     struct FixedMatcher(Vec<((usize, usize), u32)>);
     impl PipelinePatternMatcher for FixedMatcher {
